+++ clients/adapter/src/main/java/org/eclipse/hono/adapter/client/command/CommandRouterCommandConsumerFactory.java
@@ -24,6 +24,7 @@ import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
 import org.eclipse.hono.client.ClientErrorException;
+import org.eclipse.hono.client.ConnectionLifecycle;
 import org.eclipse.hono.client.ServiceInvocationException;
 import org.eclipse.hono.util.Lifecycle;
 import org.eclipse.hono.util.Strings;
@@ -54,6 +55,8 @@ public class CommandRouterCommandConsumerFactory implements CommandConsumerFacto
     private final CommandRouterClient commandRouterClient;
     private final List<Lifecycle> internalCommandConsumers = new ArrayList<>();
 
+    private int maxTenantIdsPerRequest = 100;
+
     /**
      * Creates a new factory.
      *
@@ -64,6 +67,30 @@ public class CommandRouterCommandConsumerFactory implements CommandConsumerFacto
     public CommandRouterCommandConsumerFactory(final CommandRouterClient commandRouterClient, final String adapterName) {
         this.commandRouterClient = Objects.requireNonNull(commandRouterClient);
         this.adapterInstanceId = getNewAdapterInstanceId(Objects.requireNonNull(adapterName));
+        if (commandRouterClient instanceof ConnectionLifecycle<?>) {
+            ((ConnectionLifecycle<?>) commandRouterClient).addReconnectListener(con -> reenableCommandRouting());
+        }
+    }
+
+    void setMaxTenantIdsPerRequest(final int count) {
+        this.maxTenantIdsPerRequest = count;
+    }
+
+    private void reenableCommandRouting() {
+        final List<String> tenantIds = commandHandlers.getCommandHandlers().stream()
+                .map(CommandHandlerWrapper::getTenantId)
+                .distinct()
+                .collect(Collectors.toList());
+
+        int idx = 0;
+        // re-enable routing of commands in chunks of tenant IDs
+        while (idx < tenantIds.size()) {
+            final int from = idx;
+            final int to = from + Math.min(maxTenantIdsPerRequest, tenantIds.size() - idx);
+            final List<String> chunk = tenantIds.subList(from, to);
+            commandRouterClient.enableCommandRouting(chunk, null);
+            idx = to;
+        }
     }
 
     private static String getNewAdapterInstanceId(final String adapterName) {
@@ -86,7 +113,9 @@ public class CommandRouterCommandConsumerFactory implements CommandConsumerFacto
      */
     public void registerInternalCommandConsumer(
             final BiFunction<String, CommandHandlers, Lifecycle> internalCommandConsumerSupplier) {
-        internalCommandConsumers.add(internalCommandConsumerSupplier.apply(adapterInstanceId, commandHandlers));
+        final Lifecycle consumer = internalCommandConsumerSupplier.apply(adapterInstanceId, commandHandlers);
+        log.info("register internal command consumer {}", consumer.getClass().getSimpleName());
+        internalCommandConsumers.add(consumer);
     }
 
     /**
@@ -118,9 +147,6 @@ public class CommandRouterCommandConsumerFactory implements CommandConsumerFacto
         return CompositeFuture.all(futures).mapEmpty();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public final Future<CommandConsumer> createCommandConsumer(
             final String tenantId,
@@ -136,9 +162,6 @@ public class CommandRouterCommandConsumerFactory implements CommandConsumerFacto
         return doCreateCommandConsumer(tenantId, deviceId, null, commandHandler, lifespan, context);
     }
 
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public final Future<CommandConsumer> createCommandConsumer(
             final String tenantId,